refactor(tui::ui::components::popups): move "HelpPopup" to own file
